# backend/src/main/java/com/greenloop/auth/ErrorResponse.java
package com.greenloop.auth;

import com.fasterxml.jackson.annotation.JsonFormat;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

/**
 * Standardized error response DTO for all API endpoints.
 * Provides consistent error information across the application.
 */
@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class ErrorResponse {

    /**
     * Timestamp when the error occurred.
     */
    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
    private LocalDateTime timestamp;

    /**
     * HTTP status code.
     */
    private int status;

    /**
     * Error type/category (e.g., "Unauthorized", "Token Expired").
     */
    private String error;

    /**
     * Detailed error message.
     */
    private String message;

    /**
     * API path where the error occurred.
     */
    private String path;
}