# openapi.yaml
openapi: 3.1.0
info:
  title: GreenLoop Food Rescue API
  description: |
    Comprehensive API for the GreenLoop food rescue application.
    CS 520 Spring 2026, Team 12.

    The GreenLoop platform connects food retailers, dining halls, and donors with consumers
    to rescue surplus food, reduce waste, and build community. Users can be Consumers,
    Retailers, DiningHalls, or Donors.
  version: 1.0.0
  contact:
    name: GreenLoop Support
    email: support@greenloop.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.greenloop.app/v1
    description: Production server
  - url: https://staging-api.greenloop.app/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from /auth/login or /auth/refresh.
        Include in Authorization header: "Authorization: Bearer <token>"
  schemas:
...
    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Email verification token from email link
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    # Listing schemas
...
    PickupSlot:
      type: object
      required:
        - start
        - end
        - reserved
      properties:
        start:
          type: string
          format: date-time
          example: "2026-04-09T17:00:00Z"
        end:
          type: string
          format: date-time
          example: "2026-04-09T18:00:00Z"
        reserved:
          type: boolean
          example: false
...
    Reservation:
      type: object
      required:
        - id
        - listing
        - consumer
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "770e8400-e29b-41d4-a716-446655440002"
        listing:
          type: object
          required:
            - id
            - title
            - owner
          properties:
            id:
              type: string
              format: uuid
            title:
              type: string
            owner:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                name:
                  type: string
        consumer:
          type: object
          required:
            - id
            - name
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            email:
              type: string
              format: email
            phone:
              type: string
        pickupSlot:
          $ref: "#/components/schemas/PickupSlot"
        status:
          type: string
          enum: [Pending, Confirmed, Completed, Cancelled, NoShow]
          example: "Confirmed"
        verificationCode:
          type: string
          description: QR code or numeric code for verification
          example: "GLP-A4B9-X2K7"
        verifiedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
          example: "2026-04-09T11:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-04-09T15:45:00Z"
    CreateReservationRequest:
      type: object
      required:
        - listingId
        - pickupSlotIndex
      properties:
        listingId:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        pickupSlotIndex:
          type: integer
          minimum: 0
          example: 0
...
    CancelReservationRequest:
      type: object
      properties:
        reason:
          type: string
          example: "Unable to make pickup time"
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Email is required"
        code:
          type: string
          example: "INVALID_INPUT"
        details:
          type: object
          description: Additional error details
          example:
            field: "email"
            reason: "Invalid email format"

    # Dashboard schemas
...
paths:
...
  /auth/verify-email:
    post:
      tags:
        - Authentication
      summary: Verify email address
      description: Verify email using token sent to user's inbox
      operationId: verifyEmail
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
      responses:
        200:
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email verified successfully"
        400:
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /reservations:
    post:
      tags:
        - Reservations
      summary: Create reservation
      description: Reserve a food listing for pickup
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationRequest"
      responses:
        201:
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden - must be Consumer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Listing or slot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: Conflict - slot already reserved or listing unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /reservations/{id}/cancel:
    patch:
      tags:
        - Reservations
      summary: Cancel reservation
      description: Cancel a reservation (consumer or listing owner)
      operationId: cancelReservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelReservationRequest"
      responses:
        200:
          description: Reservation cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        403:
          description: Forbidden - cannot cancel this reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Reservation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...